+++ cmd/vshell/init_other.go
@@ -2,6 +2,11 @@
 
 package main
 
+import (
+	"errors"
+	"time"
+)
+
 // initMain is never reached: only Linux has a PID 1 to be.
 func initMain() {}
 
@@ -14,3 +19,5 @@ func modprobe([]string) int { return 1 }
 func saveKernelLog() {}
 
 func wifiDiagnosis() string { return "" }
+
+func setSystemClock(time.Time) error { return errors.New("only the console sets its clock") }